Align Order status FK type with Status primary key

+++ Caskr.Server.Tests/OrdersServiceTests.cs
@@ -20,7 +20,7 @@ public class OrdersServiceTests
     [Fact]
     public async Task GetOrdersAsync_ReturnsRepositoryResults()
     {
-        var expected = new[] { new Order { Id = 1, StatusId = StatusType.ResearchAndDevelopment } };
+        var expected = new[] { new Order { Id = 1, StatusId = (int)StatusType.ResearchAndDevelopment } };
         _repo.Setup(r => r.GetOrdersAsync()).ReturnsAsync(expected);
 
         var result = await _service.GetOrdersAsync();
@@ -31,7 +31,7 @@ public class OrdersServiceTests
     [Fact]
     public async Task GetOrderAsync_DelegatesToRepository()
     {
-        var expected = new Order { Id = 2, StatusId = StatusType.AssetCreation };
+        var expected = new Order { Id = 2, StatusId = (int)StatusType.AssetCreation };
         _repo.Setup(r => r.GetOrderAsync(2)).ReturnsAsync(expected);
 
         var result = await _service.GetOrderAsync(2);
@@ -42,7 +42,7 @@ public class OrdersServiceTests
     [Fact]
     public async Task AddOrderAsync_DelegatesToRepository()
     {
-        var order = new Order { Id = 3, StatusId = StatusType.TtbApproval };
+        var order = new Order { Id = 3, StatusId = (int)StatusType.TtbApproval };
         _repo.Setup(r => r.AddOrderAsync(order)).ReturnsAsync(order);
 
         var result = await _service.AddOrderAsync(order);
@@ -53,8 +53,8 @@ public class OrdersServiceTests
     [Fact]
     public async Task UpdateOrderAsync_DelegatesToRepository()
     {
-        var order = new Order { Id = 4, OwnerId = 1, StatusId = StatusType.Ordering };
-        _repo.Setup(r => r.GetOrderAsync(order.Id)).ReturnsAsync(new Order { Id = 4, OwnerId = 1, StatusId = StatusType.AssetCreation });
+        var order = new Order { Id = 4, OwnerId = 1, StatusId = (int)StatusType.Ordering };
+        _repo.Setup(r => r.GetOrderAsync(order.Id)).ReturnsAsync(new Order { Id = 4, OwnerId = 1, StatusId = (int)StatusType.AssetCreation });
         _repo.Setup(r => r.UpdateOrderAsync(order)).ReturnsAsync(order);
 
         var result = await _service.UpdateOrderAsync(order);
@@ -66,8 +66,8 @@ public class OrdersServiceTests
     [Fact]
     public async Task UpdateOrderAsync_StatusChangedToTtbApproval_SendsEmail()
     {
-        var order = new Order { Id = 6, OwnerId = 2, StatusId = StatusType.TtbApproval, Name = "Order" };
-        _repo.Setup(r => r.GetOrderAsync(order.Id)).ReturnsAsync(new Order { Id = 6, OwnerId = 2, StatusId = StatusType.AssetCreation });
+        var order = new Order { Id = 6, OwnerId = 2, StatusId = (int)StatusType.TtbApproval, Name = "Order" };
+        _repo.Setup(r => r.GetOrderAsync(order.Id)).ReturnsAsync(new Order { Id = 6, OwnerId = 2, StatusId = (int)StatusType.AssetCreation });
         _repo.Setup(r => r.UpdateOrderAsync(order)).ReturnsAsync(order);
         var user = new User { Id = 2, Email = "owner@example.com", Name = "Owner" };
         _usersRepo.Setup(r => r.GetUserAsync(order.OwnerId)).ReturnsAsync(user);
